[nan-1643] dont use bash script

+++ tests/integration-templates/generate-and-run-tests.bash
@@ -9,14 +9,17 @@ popd () {
 
 INTEGRATION=$1
 
+if [ ! -d "../../integration-templates/$INTEGRATION/mocks" ]; then
+    echo "No mocks found for $INTEGRATION"
+    exit 0
+fi
+
 mkdir -p nango-integrations
 pushd nango-integrations
 cp -r ../../../integration-templates/$INTEGRATION .
 mv $INTEGRATION/nango.yaml .
 npx nango generate
 npm run generate
-pushd ../tests
-popd
 npm run test
 popd
 rm -rf nango-integrations
